Rename namespace App\Events to App\Event to be in sync with the docs
